refactor: moving AuthContext from 'src\components\EditPage\temp_data\trip_reducer.jsx' to 'src\contexts\AuthContext.jsx'
